+++ style_bert_vits2/models/modules.py
@@ -5,10 +5,10 @@ from torch import nn
 from torch.nn import Conv1d
 from torch.nn import functional as F
 from torch.nn.utils import remove_weight_norm, weight_norm
-from transforms import piecewise_rational_quadratic_transform
 
 from style_bert_vits2.models import commons
 from style_bert_vits2.models.attentions import Encoder
+from style_bert_vits2.models.transforms import piecewise_rational_quadratic_transform
 
 
 LRELU_SLOPE = 0.1
